[OING-105] feat: 캘린더 API를 위한 테스트 추가 (#104)
* feat: Add the integration test for weekly calendarApiTest

* feat: Add edge case of the integration test at CalendarApiTest

* feat: Add unit test for CalendarController

* feat: Add DataJpaTest for MemberPostRepositoryCustom methods for calendar

* refactor: Change variable name from User to Member

+++ gateway/src/main/java/com/oing/restapi/CalendarApi.java
@@ -33,7 +33,7 @@ public interface CalendarApi {
 
             @RequestParam(required = false)
             @Parameter(description = "조회할 주차", example = "1")
-            Long week
+            Integer week
     );
 
     @Operation(summary = "월별 캘린더 조회", description = "월별 캘린더를 조회합니다.")
